Added Combos

+++ include/utils/spring.h
@@ -3,8 +3,8 @@
 typedef struct Spring
 {
     int restLength;
-    float damping;
-    float velocity;
-    float force;
-    float k;
+    double damping;
+    double velocity;
+    double force;
+    double k;
 } Spring;
